Offer Letter Module

+++ database/migrations/2026_03_16_084622_create_final_fees_table.php
@@ -27,7 +27,7 @@ return new class extends Migration
             $table->decimal('semesterEnrollFee', 13, 2);
             $table->decimal('taxFee', 13, 2);
             $table->decimal('service_charge', 13, 2);
-            $table->decimal('credit_hour', 13, 2);
+            $table->integer('credit_hour');
             $table->decimal('per_credit_hour', 13, 2);
             $table->decimal('total_fee', 13, 2);
             $table->decimal('income_tax', 13, 2);
